openai-evaluation-streamlit/streamlit-app/requirements.txt
streamlit==1.38.0
pandas==2.2.2
requests
matplotlib==3.9.2
python-dotenv==1.0.1
